Creation of test_new_method_tracks.py
To find the maximum in the graph of the hough transform, a grid was used, and the surface occupied by each hit is incremented on the grid. This method has proven to be even slower than the original one, and should thus be ignored, unless a significant optimisation can be found.