Update test biais hough transform
The file test_artificial_tracks.py was renamed to artificial_tracks.py

The file test_new_method_tracks.py was renamed to other_hough_transform.py

The function create_artificial_tracks was updated to deal with the cases where the track goes through the corners of 4 cells.

The function new_method_tracks was corrected.

The previous hough transform function was added to the file other_hough_transform.py as old_method_tracks and was adapted to the new grid convention