Finding muon decay and computing lifetime
The function dist_line_pt in utils/physics.py was corrected

The function plot_hits in tracking/track_reconstruction.py was adapted to be able to plot hits of the next event in an other color

The following files were created in a new directory "muon_lifetime"

A file find_muon_decay.ipynb was created where the events candidate for muon decay are found, and the time between the muon track and its decay is computed for each candidate, then stored.

A file compute_lifetime.ipynb was created were the time intervals between muon track and decay are loaded, turned into a histogram. The histogram is then fitted to find the life time