Change of directory name
The directory muon_lifetime was renamed to muon_decay.

In the file find_muon_decay, the data of the hits and track of muon decay are now saved as pickles containing all the candidates events for one dataset.

The file analyse_decay_data.ipynb was created to analyse the data saved as pickle mentionned above.