zfit finalization, code improvement through functions
The file "analyse_decay_data.ipynb" was adapted to make an histogram of the energy deposit of the electrons for all the available data.

In "compute_lifetime.ipynb", the zfit function was completed (chisquare_manual function added), the plot now shows the error on the fitted lifetime.

The filtering process to find candidates for muon decays was put in a unique function in a new file "find_muon_decay.py" and the original "find_muon_decay.ipynb" was renamed to "raw_data_analysis.ipynb". The code needed to load the data of a fun was also put in a function "load_run" in "data_loading.py".

The code of "test_hough_transforms.ipynb" was completed to study the error on the angle (and not only the tangent).

The code of "test_kalman.ipynb" was adapted to load correctly the new data.

In "track.py", the kalman filter was corrected : the measure uncertainty due to the size of the cells is put in the matrix R insteak of P. P is now an identity matrix.